stats and feature flipping

+++ back/controllers/FeatureFlippedController.go
@@ -3,13 +3,30 @@ package controllers
 import (
 	"backend/initializers"
 	"backend/models"
+	roleCheck "backend/utils"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
-// FeaturesList - List all features
+// FeaturesList godoc
+// @Summary List features
+// @Description Retrieves a list of all feature toggles
+// @Tags features
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Success 200 {array} swaggermodels.FeatureFlippedSwagger "List of features"
+// @Failure 401 {string} string "Unauthorized if not admin"
+// @Failure 500 {string} string "Internal server error on fetching features"
+// @Router /features [get]
 func FeaturesList(c *gin.Context) {
+
+	if !roleCheck.IsAdmin(c) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	var features []models.FeatureFlipped
 	result := initializers.DB.Find(&features)
 	if result.Error != nil {
@@ -19,8 +36,25 @@ func FeaturesList(c *gin.Context) {
 	c.JSON(http.StatusOK, features)
 }
 
-// FeatureCreate - Create a feature
+// FeatureCreate godoc
+// @Summary Create feature
+// @Description Creates a new feature toggle
+// @Tags features
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Param feature body swaggermodels.FeatureFlippedSwagger true "Feature to create"
+// @Success 201 {object} swaggermodels.FeatureFlippedSwagger "Feature created"
+// @Failure 400 {string} string "Bad request if data is incorrect"
+// @Failure 401 {string} string "Unauthorized if not admin"
+// @Failure 500 {string} string "Internal server error on creating feature"
+// @Router /features [post]
 func FeatureCreate(c *gin.Context) {
+	if !roleCheck.IsAdmin(c) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	var feature models.FeatureFlipped
 	if err := c.ShouldBindJSON(&feature); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -34,8 +68,24 @@ func FeatureCreate(c *gin.Context) {
 	c.JSON(http.StatusCreated, feature)
 }
 
-// FeatureGet - Get a single feature
+// FeatureGet godoc
+// @Summary Get feature
+// @Description Retrieves a specific feature by ID
+// @Tags features
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Param id path string true "Feature ID"
+// @Success 200 {object} swaggermodels.FeatureFlippedSwagger "Feature retrieved"
+// @Failure 404 {string} string "Feature not found"
+// @Failure 401 {string} string "Unauthorized if not admin"
+// @Router /features/{id} [get]
 func FeatureGet(c *gin.Context) {
+	if !roleCheck.IsAdmin(c) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	var feature models.FeatureFlipped
 	if err := initializers.DB.Where("id = ?", c.Param("id")).First(&feature).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{"error": "Feature not found"})
@@ -44,8 +94,26 @@ func FeatureGet(c *gin.Context) {
 	c.JSON(http.StatusOK, feature)
 }
 
-// FeatureUpdate - Update a feature
+// FeatureUpdate godoc
+// @Summary Update feature
+// @Description Updates an existing feature toggle
+// @Tags features
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Param id path string true "Feature ID"
+// @Param feature body swaggermodels.FeatureFlippedSwagger true "Feature data to update"
+// @Success 200 {object} swaggermodels.FeatureFlippedSwagger "Feature updated"
+// @Failure 400 {string} string "Bad request if data is incorrect"
+// @Failure 404 {string} string "Feature not found"
+// @Failure 401 {string} string "Unauthorized if not admin"
+// @Router /features/{id} [patch]
 func FeatureUpdate(c *gin.Context) {
+	if !roleCheck.IsAdmin(c) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	var feature models.FeatureFlipped
 	if err := initializers.DB.Where("id = ?", c.Param("id")).First(&feature).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{"error": "Feature not found"})
@@ -61,8 +129,25 @@ func FeatureUpdate(c *gin.Context) {
 	c.JSON(http.StatusOK, feature)
 }
 
-// FeatureDelete - Delete a feature
+// FeatureDelete godoc
+// @Summary Delete feature
+// @Description Deletes a specific feature toggle
+// @Tags features
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Param id path string true "Feature ID"
+// @Success 200 {string} string "Feature deleted"
+// @Failure 404 {string} string "Feature not found"
+// @Failure 401 {string} string "Unauthorized if not admin"
+// @Failure 500 {string} string "Internal server error on deleting feature"
+// @Router /features/{id} [delete]
 func FeatureDelete(c *gin.Context) {
+	if !roleCheck.IsAdmin(c) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	if err := initializers.DB.Delete(&models.FeatureFlipped{}, c.Param("id")).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
